First Complete Commit for CMU-RoadWork and Mapillary-Vistas processing files
