--- DomainSeg/create_masks/mapillary_vistas_2.0/.ipynb_checkpoints/README-checkpoint.md
@@ -1,3 +1,0 @@
-# DomainSeg Network
-
-The DomainSeg network is part of the AutoSeg vision foundation model and is responsible for construction zone objects segmentation.
