--- DomainSeg/create_masks/mapillary_vistas_2.0/val_address_modified.csv
@@ -1,185 +1,0 @@
-address
-/-9y4NjcjdoPFMs5wwC7otg
-/-BYnT4s40fJHAlOumPYbyQ
-/-F5vhdPopdHyJjiC2hI6xg
-/-YfOWQM-geslv541gdpQPQ
-/0wcA84wys1Ag1X-_L8PreA
-/1S5dk82Kz6LuXJed4xXfqA
-/2DPJuTp0mSTXXGEsWRL25g
-/2RfwaKn8rVSFH2epy9rJqQ
-/3dDUDmy9HoIFzIJtmt-V9Q
-/3GtKr4T8GuqDgGpIS5TRKg
-/3mDvhJz0Te7yryhj3Z970g
-/3mz6ZPAfbxapSa0j3klqjg
-/3Ned3pRUlZ5NcT9qP5b_BA
-/3_f1Ewum_g-gdZ9h-SuKKA
-/42iZxKOcI70Ab-UZ8RNdUg
-/4iq7AAv9yhZaTAoOIJVorQ
-/4QTysV22Ql4Q7ginGb2HLw
-/4XNyiO49GE1W6chmXeP3ig
-/57RX6_BDC8sYop4gU8DpUw
-/5K1_i38r46HrvLTfSPn_TQ
-/5S-ns93ghhPGmmS8HRUDtg
-/63B8VdNHii23jy9_xCNCXw
-/6AmjQLaXDIbwaFH7WJl3Mw
-/6I8hQ7QwJrLsSEjSZnjYzw
-/6KP-F9snLfi2KtIuDbOYAg
-/6MPIkgIvaK8gQwonURcdeQ
-/6UaGqFikn00j8xkfYf-goQ
-/6VjCwuYEIN3NG-LHhoCOGg
-/6x2rKAIi42pIXowhKP20Xw
-/7o-Zx6Y4-T48JQPwBN7gbg
-/7sfF1oLkgj7MX8mQclJncA
-/8anx4_eW_MnMjbJcTYSM8A
-/8FF58wpNZtOO8WjdFASPcg
-/94lRvRGNBLZIbfcgUmcpRQ
-/9vmLvh9pkmsviZfqVuws-Q
-/9wQDralITa7o-t-ac3_32Q
-/a2-cjwuXmO687YN_vlynhw
-/a4AAMxVgsdb7XpBd6x7PRg
-/aAcaQrZy91einjB0af8YFg
-/AEodz50qZwUPrCsA5tpWig
-/Afk5LVg5hXl3ITK8uBvtjg
-/AGxQK4GN_qg5qCl-e0KpWQ
-/aN9zpqHrwNqpbtoO9ENuZA
-/AyMB6Add5VPGOCf4NjcEgQ
-/AYWRYnH4_RSTvQ9gL4_3gg
-/AzpT2miuJpghiLLEikt1og
-/b8LkVdta0t-0IZ5rBRe1Qw
-/b91fZvSFEVfHTOFou8pHFQ
-/bbfNQHuYLehPra3l-WZJRg
-/bb_Qd6uhG9Puk8Z3mWnkwQ
-/bdFCJ8-cpipNckz6rXKH-A
-/bdVw3ht5no3Itw5eQVSpzQ
-/bEc2b8MaP6GzNWfUks7gwQ
-/CoBAkHgeaj5djbvAhayQ1g
-/cq-e26vIeu1ZscriUqpJVA
-/CYdn7spjYQGeczeQh3nugQ
-/dhMr__VcDjJHTgtLrfMZtQ
-/DL6RBxuTxAlgVVdAucmd4g
-/dm4y-sOUvmTy4VcAjPuIpA
-/dWDcQMFCAXHcy8yGl2qFLQ
-/DX6nJZ-lB4o8OofJlb15PQ
-/E00oWA0zsmrvpDmZGxNw_w
-/EBoMwNOLsxUGoVEmEc3AcA
-/EfNkxlj-XjXAnNNJZqbz5Q
-/eGre90jefz7x9b_OxlM4nw
-/ejvvpHQzMIGEScb9nn-UVQ
-/EwHYWvWqJ82jIQr7B0sUcw
-/F-Zk_X9-a84Tom14YCgdcQ
-/fES_Q-LpcKlmJmK9PxlUdA
-/Fimn6C65kmRy1bO8BdLQzg
-/FIPQ_-byAFERw3FIZMQArA
-/FyC2cimDdAGyRbfhFCr-7g
-/FYoB0UGu40k9nKurB8pucw
-/gFfZ5yDbjysz5DWiyT1-5Q
-/GiwWuQKUYVpMSMg_-FN1jQ
-/Gp0KqhB88_8Jbtarq0f1Dg
-/gxv9RvgHZi_XF11EbJ4Opw
-/hf2QGmB3ULp0yw7WanicuA
-/HMPHvE04RN1c21pdN2XUSw
-/Hngs3e9MUJscCPUCMuXODQ
-/HwBonTfmkfIOANA1O7B2OQ
-/I2PhN7ftmK46zYSCDA-_Xw
-/I5KCGCImogJHoMEuss4Opw
-/IBF4NeAjWS64aDsjFRVZzw
-/iBQN6CKfcMTNpXGhr5-69w
-/Ik5C23HH5aeFdc311WYdeg
-/IOXKHTTt-dAgrVtMXsUGQA
-/irTuEwhCHDez_innQChlNw
-/iYGUOkS9D57bpnPOLgf_7A
-/Iz8irvLYSOcDrIbWGeKlAQ
-/J5BFlgXZdSRc_WAWjs8JHw
-/jEUk_wzXO_ft8gCIPp9ddA
-/JSCTB43EcE3jcrH2kZk8eA
-/jTHuwAuMdl-50usve0zP_A
-/jzW7IO0aja2Ww9aQtiag7Q
-/k9UCSxm_NtkoQRVNp4Mn2A
-/kCA4mZhQTMahGWqLhh4c1g
-/klV2LGDKxRSj37HA_n8eeA
-/KnM8yd2cabsAlGqxWtCYXA
-/KSq23J1ibgw_ufqayMOGPA
-/l2Zi8r7JRWoiS2vjcfHQgg
-/lIYCcBPlNAhb_kUb9dGhAA
-/LK5D2QOplPLAmkn2GghgOw
-/lKbf3ZUrNlZtTHkPaVs5kQ
-/LydCIsQfkzyJ5TLSSBNm5w
-/matatT_eRHwLZcJhUdE7CQ
-/mDO_3Xlc2q0OB1nNGXy_Gw
-/ME5Y_enkRvyMK940tDeWTw
-/MyusQnoLP84Wv3sgOgJ3IA
-/N6JH6pMYosRTHjMT_JOMKg
-/Nf8J6O9-SL7J_TrvbMsGew
-/NxPtGt032tyZF2WrZXjOzw
-/OJCZor7bH2DWkdTM2lrAiA
-/or6qWEt_MNF83_IWIY3e8g
-/oWCF3yBXX-NmmuEl6Oa1YA
-/OYyFv3XcyrSla0sgF6JrEg
-/P9IpY9o3B02j1P7iutiE4g
-/PHzptJeal8eLwpX08uw4aQ
-/PIwuiRK5KLmLYCwIf1jPTw
-/PLn2SwsrAXovfYeQrA9T4Q
-/pO28OyoV7uhea8_epCywlw
-/pq3yl1U99I-Zqhk_d2vUKw
-/pRD2yHe-TrhydDf-k1oDSA
-/pwIX_FS2F4Z85rqpeGXQHg
-/PxmqFeA1n_avfakO17JaZg
-/qg9BNe9SLUwa7zBbHbigsQ
-/QPc3tzYRQZ3h6z_9VJ8LXw
-/qujdQhR-JNNXuHmAD_GeMA
-/QxzH_y5v6wuolpo8qLc2VA
-/R9eSYGfxiuF9YCrH9PNRLA
-/rgaUarx1B5HnNcLC94fGeQ
-/rgfq0PJbbWd3XoPI7f2iQg
-/RhJoCJV_kj2B4ctzhP03vw
-/RLCQeZoyi1qeimRxMUZRdA
-/RvHEAMAjsiXe3QUcrHmw4A
-/rX4E0IJIT-I1uWZAedh8Yw
-/s9pp8kBaUxg0KMGDSgIZiA
-/SqBZ2HitJVan9f9yrXPGWA
-/stuxhvfBJixSEWtztbUK0A
-/SYfe6CtkLAAVhZjHBFTtPQ
-/TckCiT3txFGhUa_-8-KcbQ
-/TGOxdkO7wSqOjFD06v_RVg
-/TLeRS2WxTGPLGwLFc_sIeA
-/Ts7ygZkVTPclnNqHsSWn-w
-/TSQLBOq8vUh6IfkJC6BbfA
-/tWyD7pJUk60MwWdlFABkoA
-/U4mBEuqtEQ1EKB5RHfWNhQ
-/uGlZUb9e1jZuFxUYluhcRQ
-/uMwBwpfHHKQ6so9ljmWn8Q
-/v2XSpSateiSSoPeKfsX32g
-/VaNDyP50fMsQTvj3Usufig
-/vAwEpBbiIuer1puNj1RH3Q
-/VMpwgzw0t-SMq28xvcdXOg
-/VT1d62Q_MLlNsDWgBxvGNg
-/w6A5-J0EYAAuOyGWhPL1iA
-/wer8bUy6kZUILZEz1TjNRw
-/wfD7PJRqBfh9Nc4yDp_GvQ
-/wj-UHOWQS12Y2XxIPvqc2A
-/Wk5YC67h46rl2P3hrZKe7A
-/wMveRkPBCUlgFdy-AtQwbQ
-/wVKyMdCkg2mLGrmlh2wEmw
-/WvTCcXp__xX4yj_JBwaJtw
-/wWKUDfNfxDXzHtlTCKQXeA
-/x0mvKia_FhNfXNgtHnrqRA
-/x7zP3tETzGUbJO9CSoqrXg
-/xepZVIlkU1tPmMy6BkH7WQ
-/XGHvbPqeWY49EsxQqyR-TA
-/xPRlO9PAG7cIWNICs5hDwA
-/xTy7q_hU6_14KyxbDrJj1w
-/YHywwM1X2Vkj-2G1zp9x5Q
-/YnRehAQXiEfX2VgaFLGwBg
-/yQnWvJVWVvfwguvdPqm74Q
-/YQonW9awgLsEX5RAgShLQg
-/YrUVWEcCBvU5n2oDBNvcSQ
-/z7s0jAjdDVUbOk8QfYK9dA
-/ZbL1-7zekkF-7ABgsBu_Fw
-/Zga1j2Cc7ED8TuPEdQTV4Q
-/zIa42cgedZ6Elyn6ty4Vxg
-/ZP4QyBO7v7dwcNTKrJ-rOA
-/ZYG642weurHoJ2aKEszS-g
-/_1Gn_xkw7sa_i9GU4mkxxQ
-/_EGt7GTUW6OKw0zzz7c5-A
-/_gcgraos355QHyu7iFMK1w
-/_zp7_itt7iZM5aR9lFCLyg
